pyproject.toml
[project]
name = "scam-honeypot-ai"
version = "0.1.0"
description = "Scam Honeypot AI System"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.100.0",
    "uvicorn==0.24.0",
    "pydantic==1.10.20",
    "python-dotenv==1.0.0",
    "google-generativeai==0.3.0",
    "httpx==0.24.0",
    "requests==2.31.0",
]
